+++ src/main/java/com/example/p2b/service/PreferService.java
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class PreferService {
     private final PreferRepository preferRepository;
     @Transactional
-    public void addPrefer(String pdname, int pdtype, String pdaddr, String pdtel, int pdlocal, float pdpoint, int pdprice, int pdwedprice, User user){
+    public void addPrefer(String pdname, int pdtype, String pdaddr, String pdtel, int pdlocal, double pdpoint, int pdprice, int pdwedprice, User user){
         Prefer prefer = new Prefer(pdname, pdtype, pdaddr, pdtel, pdlocal, pdpoint, pdprice, pdwedprice, user);
         preferRepository.save(prefer);
     }
